requirements.txt
flask
aiohttp
python-dotenv